Wiring everything together and passing first acceptance test (caveat that currency display is currently the same as returned by the api instead of the currency symbol. fixing later)

+++ app/src/androidTest/java/com/savvasdalkitsis/mondo/test/ui/actors/assertions/UserAssertions.java
@@ -6,7 +6,7 @@ import static com.savvasdalkitsis.mondo.test.assertions.SeesAssertions.seesText;
 
 public class UserAssertions {
 
-    public void seesBalance(int balance, TestCurrency currency) {
+    public void seesBalance(double balance, TestCurrency currency) {
         seesText("Balance");
         seesText(String.valueOf(balance));
         seesText(currency.displayString());
